Use kcrypt lib directly for encrypt
Also move the uki kcrypt hook into its own file
